requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
mediapipe>=0.10.0
anthropic>=0.5.0  # For Claude API
PyQt6>=6.5.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0 